Export find IPReservation matching ID to a func
Currently, matching the owner of the IP when deallocating the
address happens only based on the containerID.

In follow-up commits we will require this behavior to happen based
on the pod reference attribute.

This commit prepares the code base for that, by enabling the caller
to inject the criteria for identifying the owner of an IP into the
IP address de-allocation function.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>